refactor: Product JPA 전환
