+++ src/main/java/kitchenpos/application/MenuService.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.dao.MenuProductDao;
-import kitchenpos.dao.ProductDao;
+import kitchenpos.dao.ProductRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.Product;
@@ -26,18 +26,18 @@ public class MenuService {
     private final MenuDao menuDao;
     private final MenuGroupDao menuGroupDao;
     private final MenuProductDao menuProductDao;
-    private final ProductDao productDao;
+    private final ProductRepository productRepository;
 
     public MenuService(
         MenuDao menuDao,
         MenuGroupDao menuGroupDao,
         MenuProductDao menuProductDao,
-        ProductDao productDao
+        ProductRepository productRepository
     ) {
         this.menuDao = menuDao;
         this.menuGroupDao = menuGroupDao;
         this.menuProductDao = menuProductDao;
-        this.productDao = productDao;
+        this.productRepository = productRepository;
     }
 
     @Transactional
@@ -80,7 +80,7 @@ public class MenuService {
     }
 
     private Product getProductById(Long productId) {
-        return productDao.findById(productId)
+        return productRepository.findById(productId)
             .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 상품입니다."));
     }
