+++ src/main/java/kitchenpos/dao/ProductRepository.java
@@ -1,0 +1,11 @@
+package kitchenpos.dao;
+
+import kitchenpos.domain.Product;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
