+++ src/main/java/kitchenpos/dao/jdbc/JdbcTemplateMenuGroupDao.java
@@ -1,6 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.dao.jdbc;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.dao.MenuGroupDao;
+import kitchenpos.domain.MenuGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -12,19 +13,18 @@ import org.springframework.stereotype.Repository;
 import javax.sql.DataSource;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public class JdbcTemplateTableGroupDao implements TableGroupDao {
-    private static final String TABLE_NAME = "table_group";
+public class JdbcTemplateMenuGroupDao implements MenuGroupDao {
+    private static final String TABLE_NAME = "menu_group";
     private static final String KEY_COLUMN_NAME = "id";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public JdbcTemplateTableGroupDao(final DataSource dataSource) {
+    public JdbcTemplateMenuGroupDao(final DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -33,14 +33,14 @@ public class JdbcTemplateTableGroupDao implements TableGroupDao {
     }
 
     @Override
-    public TableGroup save(final TableGroup entity) {
+    public MenuGroup save(final MenuGroup entity) {
         final SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
         final Number key = jdbcInsert.executeAndReturnKey(parameters);
         return select(key.longValue());
     }
 
     @Override
-    public Optional<TableGroup> findById(final Long id) {
+    public Optional<MenuGroup> findById(final Long id) {
         try {
             return Optional.of(select(id));
         } catch (final EmptyResultDataAccessException e) {
@@ -49,22 +49,30 @@ public class JdbcTemplateTableGroupDao implements TableGroupDao {
     }
 
     @Override
-    public List<TableGroup> findAll() {
-        final String sql = "SELECT id, created_date FROM table_group";
+    public List<MenuGroup> findAll() {
+        final String sql = "SELECT id, name FROM menu_group";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private TableGroup select(final Long id) {
-        final String sql = "SELECT id, created_date FROM table_group WHERE id = (:id)";
+    @Override
+    public boolean existsById(final Long id) {
+        final String sql = "SELECT CASE WHEN COUNT(*) > 0 THEN TRUE ELSE FALSE END FROM menu_group WHERE id = (:id)";
+        final SqlParameterSource parameters = new MapSqlParameterSource()
+                .addValue("id", id);
+        return jdbcTemplate.queryForObject(sql, parameters, Boolean.class);
+    }
+
+    private MenuGroup select(final Long id) {
+        final String sql = "SELECT id, name FROM menu_group WHERE id = (:id)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
                 .addValue("id", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private TableGroup toEntity(final ResultSet resultSet) throws SQLException {
-        return new TableGroup(
-            resultSet.getLong(KEY_COLUMN_NAME),
-            resultSet.getObject("created_date", LocalDateTime.class)
+    private MenuGroup toEntity(final ResultSet resultSet) throws SQLException {
+        return new MenuGroup(
+            resultSet.getLong("id"),
+            resultSet.getString("name")
         );
     }
 }
