+++ src/main/java/kitchenpos/dao/jdbc/JdbcTemplateOrderDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.dao.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -17,6 +17,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
+import kitchenpos.dao.OrderDao;
 import kitchenpos.domain.Order;
 
 @Repository
