+++ src/main/resources/application.properties
@@ -1,4 +1,7 @@
 logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+
 spring.h2.console.enabled=true
+
+spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
